pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Production-style RAG system with indexing and query pipelines"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2.0",
  "python-dotenv",
  "numpy",
  "faiss-cpu",
  "sentence-transformers",
  "tqdm",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "httpx",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
